feat: add advantages links
