+++ src/components/Navbar/index.tsx
@@ -1,196 +1,187 @@
 import * as React from 'react';
-import AppBar from '@mui/material/AppBar';
+import { styled, useTheme, Theme, CSSObject } from '@mui/material/styles';
 import Box from '@mui/material/Box';
-import Toolbar from '@mui/material/Toolbar';
+import MuiDrawer from '@mui/material/Drawer';
+import MuiAppBar, { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar';
+import List from '@mui/material/List';
+import CssBaseline from '@mui/material/CssBaseline';
+import Divider from '@mui/material/Divider';
 import IconButton from '@mui/material/IconButton';
-import Typography from '@mui/material/Typography';
-import Menu from '@mui/material/Menu';
 import MenuIcon from '@mui/icons-material/Menu';
-import Container from '@mui/material/Container';
-import Avatar from '@mui/material/Avatar';
-import Button from '@mui/material/Button';
-import Tooltip from '@mui/material/Tooltip';
-import MenuItem from '@mui/material/MenuItem';
-import AdbIcon from '@mui/icons-material/Adb';
-import { Link } from 'react-router-dom';
+import ChevronRightIcon from '@mui/icons-material/ChevronRight';
+import ListItem from '@mui/material/ListItem';
+import ListItemButton from '@mui/material/ListItemButton';
+import ListItemIcon from '@mui/material/ListItemIcon';
+import ListItemText from '@mui/material/ListItemText';
+import CloseIcon from '@mui/icons-material/Close';
+import { Link} from 'react-router-dom';
+import DnsIcon from '@mui/icons-material/Dns';
+import MenuBookIcon from '@mui/icons-material/MenuBook';
+import BookIcon from '@mui/icons-material/Book';
+import SportsEsportsIcon from '@mui/icons-material/SportsEsports';
+import BarChartIcon from '@mui/icons-material/BarChart';
+import LoginIcon from '@mui/icons-material/Login';
 
-const settings = ['Sign in'];
+const drawerWidth = 240;
 
-const Navbar = () => {
-	const [anchorElNav, setAnchorElNav] = React.useState<null | HTMLElement>(
-		null 
-	);
-	const [anchorElUser, setAnchorElUser] = React.useState<null | HTMLElement>(
-		null
-	);
+const openedMixin = (theme: Theme): CSSObject => ({
+  width: drawerWidth,
+	background: '#172B4C',
+	color: 'white',
+  transition: theme.transitions.create('width', {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.enteringScreen,
+  }),
+  overflowX: 'hidden',
+});
 
-	const handleOpenNavMenu = (event: React.MouseEvent<HTMLElement>) => {
-		setAnchorElNav(event.currentTarget);
-	};
-	const handleOpenUserMenu = (event: React.MouseEvent<HTMLElement>) => {
-		setAnchorElUser(event.currentTarget);
-	};
+const closedMixin = (theme: Theme): CSSObject => ({
+  transition: theme.transitions.create('width', {
+    easing: theme.transitions.easing.sharp,
+    duration: theme.transitions.duration.leavingScreen,
+  }),
+  overflowX: 'hidden',
+  width: `calc(${theme.spacing(7)} + 1px)`,
+	background: '#172B4C',
+	color: 'white',
+  [theme.breakpoints.up('sm')]: {
+    width: `calc(${theme.spacing(8)} + 1px)`,
+  },
+});
 
-	const handleCloseNavMenu = () => {
-		setAnchorElNav(null);
-	};
+const DrawerHeader = styled('div')(({ theme }) => ({
+  display: 'flex',
+  whiteSpace: 'nowrap',
+  boxSizing: 'border-box',
+	cursor: 'pointer',
+  padding: theme.spacing(0, 1),
+  // necessary for content to be below app bar
+  ...theme.mixins.toolbar,
+}));
 
-	const handleCloseUserMenu = () => {
-		setAnchorElUser(null);
-	};
+const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(
+  ({ theme, open }) => ({
+    width: drawerWidth,
+    flexShrink: 0,
+    whiteSpace: 'nowrap',
+    boxSizing: 'border-box',
+		
+    ...(open && {
+      ...openedMixin(theme),
+      '& .MuiDrawer-paper': openedMixin(theme),
+    }),
+    ...(!open && {
+      ...closedMixin(theme),
+      '& .MuiDrawer-paper': closedMixin(theme),
+    }),
+  }),
+);
 
-	return (
-		<AppBar position='static'>
-			<Container maxWidth='xl'>
-				<Toolbar disableGutters>
-					<Typography
-						variant='h6'
-						noWrap
-						component='a'
-						href='/'
-						sx={{
-							mr: 2,
-							display: { xs: 'none', md: 'flex' },
-							fontFamily: 'monospace',
-							fontWeight: 700,
-							letterSpacing: '.3rem',
-							color: 'inherit',
-							textDecoration: 'none',
-						}}
-					>
-						RSLang
-					</Typography>
 
-					<Box sx={{ flexGrow: 1, display: { xs: 'flex', md: 'none' } }}>
-						<IconButton
-							size='large'
-							aria-label='account of current user'
-							aria-controls='menu-appbar'
-							aria-haspopup='true'
-							onClick={handleOpenNavMenu}
-							color='inherit'
-						>
-							<MenuIcon />
-						</IconButton>
-						<Menu
-							id='menu-appbar'
-							anchorEl={anchorElNav}
-							anchorOrigin={{
-								vertical: 'bottom',
-								horizontal: 'left',
-							}}
-							keepMounted
-							transformOrigin={{
-								vertical: 'top',
-								horizontal: 'left',
-							}}
-							open={Boolean(anchorElNav)}
-							onClose={handleCloseNavMenu}
-							sx={{
-								display: { xs: 'block', md: 'none' },
-							}}
-						>
-							<MenuItem onClick={handleCloseNavMenu}>
-								<Link to="/main">Main</Link>
-							</MenuItem>
-							<MenuItem onClick={handleCloseNavMenu}>
-								<Link to="/textbook">Textbook</Link>
-							</MenuItem>
-							<MenuItem onClick={handleCloseNavMenu}>
-								<Link to="/minigame">Minigame</Link>
-							</MenuItem>
-							<MenuItem onClick={handleCloseNavMenu}>
-								<Link to="/statistics">Statistics</Link>
-							</MenuItem>								
-						</Menu>
-					</Box>
-					<AdbIcon sx={{ display: { xs: 'flex', md: 'none' }, mr: 1 }} />
-					<Typography
-						variant='h5'
-						noWrap
-						component='a'
-						href=''
-						sx={{
-							mr: 2,
-							display: { xs: 'flex', md: 'none' },
-							flexGrow: 1,
-							fontFamily: 'monospace',
-							fontWeight: 700,
-							letterSpacing: '.3rem',
-							color: 'inherit',
-							textDecoration: 'none',
-						}}
-					>
-						RSLang
-					</Typography>
-					<Box sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }}>
-					<Link to="/">
-							<Button
-								onClick={handleCloseNavMenu}
-								sx={{ my: 2, color: 'white', display: 'block' }}
-							>
-								Main
-							</Button>
-							</Link>
-							<Link to="/textbook">
-							<Button
-								onClick={handleCloseNavMenu}
-								sx={{ my: 2, color: 'white', display: 'block' }}
-							>
-								Textbook
-							</Button>
-							</Link>
-							<Link to="/minigame">
-							<Button
-								onClick={handleCloseNavMenu}
-								sx={{ my: 2, color: 'white', display: 'block' }}
-							>
-								Minigames
-							</Button>
-							</Link>
-							<Link to="/statistics">
-							<Button
-								onClick={handleCloseNavMenu}
-								sx={{ my: 2, color: 'white', display: 'block' }}
-							>
-								Statistics
-							</Button>
-							</Link>
-					</Box>
 
-					<Box sx={{ flexGrow: 0 }}>
-						<Tooltip title='Open settings'>
-							<IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
-								<Avatar alt='Remy Sharp' src='/static/images/avatar/2.jpg' />
+export default function MiniDrawer(props: any) {
+  const theme = useTheme();
+  const [open, setOpen] = React.useState(false);
+
+  const handleDrawerOpen = () => {
+    setOpen(false);
+
+  };
+
+  const handleDrawerClose = () => {
+    setOpen(true);
+  };
+
+	const renderSwitch = (param: any) => {
+		switch(param) {
+			case 'Main':
+				return <DnsIcon style={{color: "white"}}/>;
+			case 'Textbook':
+				return <MenuBookIcon style={{color: "white"}} />;
+			case 'Dictionary':
+				return <BookIcon style={{color: "white"}} />;
+			case 'Minigames':
+				return <SportsEsportsIcon style={{color: "white"}} />;
+			case 'Statistics':
+				return <BarChartIcon style={{color: "white"}} />;
+			default:
+				return <DnsIcon  style={{color: "white"}}/>;
+		}
+	}
+
+  return (
+		<Box sx={{ display: 'flex', overflow: 'hidden' }}>
+			<CssBaseline />
+			<Drawer variant="permanent" open={open}>
+				<div className="navContainer" style={{height: '90vh' ,display: 'flex', flexDirection: 'column', justifyContent: 'space-between', overflow: 'hidden'}}>
+					<div className="navBox">
+							<DrawerHeader onClick={open ? handleDrawerOpen : handleDrawerClose}>
+							<IconButton
+									color="inherit"
+									aria-label="open drawer"
+								>
+								{open ? <CloseIcon/> : <MenuIcon />}
 							</IconButton>
-						</Tooltip>
-						<Menu
-							sx={{ mt: '45px' }}
-							id='menu-appbar'
-							anchorEl={anchorElUser}
-							anchorOrigin={{
-								vertical: 'top',
-								horizontal: 'right',
-							}}
-							keepMounted
-							transformOrigin={{
-								vertical: 'top',
-								horizontal: 'right',
-							}}
-							open={Boolean(anchorElUser)}
-							onClose={handleCloseUserMenu}
-						>
-							{settings.map((setting) => (
-								<MenuItem key={setting} onClick={handleCloseUserMenu}>
-									<Typography textAlign='center'>{setting}</Typography>
-								</MenuItem>
+								{theme.direction === 'rtl' ? <CloseIcon /> : <h3 style={{ opacity: open ? 1 : 0 }}>RS Lang</h3>}
+						</DrawerHeader>
+						<Divider />
+						<List>
+							{['Main', 'Textbook', 'Minigames', 'Statistics'].map((text, index) => (
+								<Link to={text.toLowerCase()}>
+									<ListItem key={text} disablePadding sx={{ display: 'block' }}>
+										<ListItemButton
+											sx={{
+												minHeight: 48,
+												justifyContent: open ? 'initial' : 'center',
+												px: 2.5,
+											}}
+										>
+											<ListItemIcon
+												sx={{
+													minWidth: 0,
+													mr: open ? 3 : 'auto',
+													justifyContent: 'center',
+												}}
+											>
+											{renderSwitch(text)}
+											</ListItemIcon>
+											<ListItemText primary={text} sx={{ opacity: open ? 1 : 0, textDecoration:"none", color: 'white' }} />
+										</ListItemButton>
+									</ListItem>
+								</Link>
 							))}
-						</Menu>
-					</Box>
-				</Toolbar>
-			</Container>
-		</AppBar>
-	);
-};
-
-export default Navbar;
+						</List>
+						<Divider />
+					</div>
+					<div className="navLog">
+					<List>
+								<Link to='login'>
+									<ListItem key='login' disablePadding sx={{ display: 'block' }}>
+										<ListItemButton
+											sx={{
+												minHeight: 48,
+												justifyContent: open ? 'initial' : 'center',
+												px: 2.5,
+											}}
+										>
+											<ListItemIcon
+												sx={{
+													minWidth: 0,
+													mr: open ? 3 : 'auto',
+													justifyContent: 'center',
+												}}
+											>
+											<LoginIcon  style={{color: "white"}}/>
+											</ListItemIcon>
+											<ListItemText primary='Sign in' sx={{ opacity: open ? 1 : 0, textDecoration:"none", color: 'white' }} />
+										</ListItemButton>
+									</ListItem>
+								</Link>
+						</List>
+					</div>	
+				</div>	
+				
+			</Drawer>
+    </Box>
+  );
+}
