+++ src/pages/Main/MainPage.model.css
@@ -50,11 +50,21 @@
 	justify-content: space-evenly;
 	flex-wrap: wrap;
 	box-shadow: 0 10px 20px rgba(0,0,0,0.4);
-
+	transition: transform 0.5s;
 }
+
+.advantages:hover{
+	transform: scale(1.1);
+}
+
 
 .advantageLogo img{
 	max-width: 80px;
+}
+
+.advantagesLink{
+	color:black;
+	text-decoration: none;
 }
 
 .videoBox{
